lotto_backend/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
sqlalchemy==2.0.25
aiosqlite==0.19.0
pydantic==2.6.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-dotenv==1.0.0
pydantic-settings==2.1.0
requests==2.31.0
email-validator==2.1.0
